Extract ignore logic to its own module.
Should allow for easier refactoring in future. I avoided adding any extra npm module dependencies that weren't already being pulled in. This may not be a worthwhile goa; I'm aware of many packages that can translate gitignore lines in to globs, of which we needn't re-invent here.
